upload-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic[email]
python-dotenv
PyJWT
python-multipart
psycopg2-binary
pika
aio_pika
